// src/app/Interfaces/objectif.interface.ts
export interface ObjectifInterface {
  idObjectif?: string;
  labelObjectif: string;
  annee: string;
  tdSeauBrute: number;
  tdSeauOsmosee: number;
  rendementMembranes: number;
  rendement: number;
  permanentFlow: number;
  tauxExploitation: number;
  deltaPMicrofiltre: number;
  deltaPMembrane: number;
  deltaPSable: number;
  cout: number;
  th: number;
  ph: number;
  tauxChlorure: number;
  nomStation: string;
}